Add datamodule for GID-15 dataset (#928)
* add datamodule with crop logic

* remove print and fix batch_size

* typo

* Use Kornia augmentations

* Style

* Ignore warning

+++ torchgeo/datasets/gid15.py
@@ -194,7 +194,7 @@ class GID15(NonGeoDataset):
             array: "np.typing.NDArray[np.int_]" = np.array(img.convert("RGB"))
             tensor = torch.from_numpy(array)
             # Convert from HxWxC to CxHxW
-            tensor = tensor.permute((2, 0, 1))
+            tensor = tensor.permute((2, 0, 1)).float()
             return tensor
 
     def _load_target(self, path: str) -> Tensor:
